Digital Design Lab Project
The goal of this project was to program a robot to receive data from a
base station using wireless asynchronous serial communication and to
pick up and drop off objects at the specified locations in the quickest,
most efficient manner possible. Pickups and drop-offs of objects are
signified by a beep and communicated to the base station.
We were provided with most of a simple computer coded in VHDL, which was
finished by the group. Things such as functions were incomplete. We were
also provided the design schematic for the robot and an incomplete UART
design schematic, as well as a safety feature for the robot. All
provided hardware and code was  provided by Dr. Thomas Collins from the
Georgia Tech Digital Design Lab course.
This project was created by Nashwin Bharwani, Greg Theodosakis, Brandi
Van De Houten, Heng Li, and Daniel Hay. I focused on the navigation
systems of the robot.
Running this code requires the Quartus II software, a Crimson Editor and
SCASM compiler and an AmigoBot outfitted with a DE2 training board and a
wireless serial communicator.